pyproject.toml
[project]
name = "raginabox"
version = "0.1.0"
description = "Visual RAG demo — see every stage of Retrieval-Augmented Generation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.45.0",
    "numpy>=2.4.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.1",
    "pymupdf>=1.25.0",
]
